refactor(remote-web): extract workspace host health query into hook
